requirements.txt
streamlit
numpy
pyvis
python-dotenv
pymongo
openai
tenacity
pathlib
typing
unidecode
torch
transformers
dataclasses
pydantic
accelerate